implement  read the pdf file and delete image path file

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const Product = require('../modles/product');
+const fileHelper = require('../util/file');
 const {validationResult} = require('express-validator/check');
 
 exports.getAddProduct = (req, res, next)=>{
@@ -143,6 +144,7 @@ exports.postEditProduct = (req, res, next) => {
            product.price = updatedPrice;
            product.description = updatedDesc;
            if(image){
+               fileHelper.deleteFile(product.imageUrl);
                product.imageUrl = image.path;
            }
            return product.save().then(result=>{
@@ -178,8 +180,13 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    //Product.findByIdAndRemove(prodId)
-    Product.deleteOne({_id: prodId, userId: req.user._id})
+    Product.findById(prodId).then(product => {
+        if(!product){
+            return next(new Error('Product not found'));
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return Product.deleteOne({_id: prodId, userId: req.user._id})
+    })    
     .then(()=>{
           console.log('DESTROYED PRODUCT');
           res.redirect('/admin/products');
@@ -188,5 +195,4 @@ exports.postDeleteProduct = (req, res, next) => {
         error.httpStatusCode = 500;
         return next(error); 
     });
-    //}).catch(err=>console.log(err));
 }
